+++ src/OpenTelemetry.Instrumentation.EventCounters/EventCountersMetrics.cs
@@ -5,7 +5,9 @@ using System.Collections.Concurrent;
 using System.Diagnostics.Metrics;
 using System.Diagnostics.Tracing;
 using System.Globalization;
-using OpenTelemetry.Internal;
+#if NET
+using System.Reflection;
+#endif
 
 namespace OpenTelemetry.Instrumentation.EventCounters;
 
@@ -31,8 +33,7 @@ internal sealed class EventCountersMetrics : EventListener
         // https://github.com/open-telemetry/opentelemetry-dotnet-contrib/issues/1024.
         _ = EventCountersInstrumentationEventSource.Log;
 
-        var assembly = typeof(EventCountersMetrics).Assembly;
-        MeterInstance = new Meter(assembly.GetName().Name, assembly.GetPackageVersion());
+        MeterInstance = Metrics.MeterFactory.Create<EventCountersMetrics>(null);
     }
 
     /// <summary>